### Player 수정 및 User 수정 , User manager 더미에서 User Scirpts로 변경

+++ Assets/Scripts/Core/interface/CommonInterface.cs
@@ -16,7 +16,7 @@ namespace Scripts.Core.inteface
     /// </summary>
     public interface IAttackable
     {
-        public int damage { get;}
+        public ulong damage { get;}
         public Vector3 attackerPos { get; }
     }
     /// <summary>
